requirements.txt
flask==3.0.0
torch
torchvision
gdown==4.7.1
pillow
numpy
pandas
pydub==0.25.1
moviepy
openai-whisper
ultralytics
opencv-python
google-genai>=1.0.0
python-dotenv>=1.0.0
